+++ libs/components/src/lib/auto-focus/auto-focus.directive.ts
@@ -1,24 +1,49 @@
 import { DOCUMENT } from '@angular/common'
-import { Directive, ElementRef, inject, OnInit } from '@angular/core'
-import { Logger, LoggerService } from '@shiftcode/ngx-core'
+import { AfterViewInit, Directive, ElementRef, Inject } from '@angular/core'
+import { isInputElement, Logger, LoggerService } from '@shiftcode/ngx-core'
 
 @Directive({ selector: '[scAutoFocus]' })
-export class AutoFocusDirective implements OnInit {
-  readonly element: HTMLElement = inject(ElementRef).nativeElement
-  private readonly document: Document = inject(DOCUMENT)
-  private readonly logger: Logger = inject(LoggerService).getInstance('AutoFocusDirective')
+export class AutoFocusDirective implements AfterViewInit {
+  readonly element: HTMLElement
+  private readonly logger: Logger
 
-  ngOnInit(): void {
-    const currentTabindex = this.element.getAttribute('tabindex')
-    if (!currentTabindex || String(parseInt(currentTabindex, 10)) !== currentTabindex) {
-      this.element.setAttribute('tabindex', '-1')
-    }
-    this.focus()
+  constructor(
+    loggerService: LoggerService,
+    elRef: ElementRef<HTMLElement>,
+    @Inject(DOCUMENT) private readonly document: Document,
+  ) {
+    this.logger = loggerService.getInstance('AutoFocusDirective')
+    this.element = elRef.nativeElement
   }
 
-  focus() {
-    this.logger.debug('try get focus on', this.element)
+  ngAfterViewInit(): void {
+    this.focus()
+    this.logger.debug('activeElement', this.document.activeElement)
+  }
+
+  focus(): boolean {
+    if (this.isDisabled() || this.isHiddenInput()) {
+      this.logger.warn('cannot focus disabled element or hidden input')
+      return false
+    }
+
+    this.logger.debug('try set focus to', this.element)
+
     this.element.focus()
-    this.logger.debug('active element:', this.document.activeElement)
+    if (this.document.activeElement !== this.element) {
+      this.element.setAttribute('tabindex', '-1')
+      this.element.focus()
+    }
+    return this.document.activeElement === this.element
+  }
+
+  private isDisabled() {
+    // using `matches` for the pseudo selector also correctly returns true when
+    // e.g. the directive is used in an input field inside a disabled fieldset
+    return this.element.hasAttribute('disabled') || this.element.matches(':disabled')
+  }
+
+  private isHiddenInput() {
+    return isInputElement(this.element) && this.element.type === 'hidden'
   }
 }