+++ libs/core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/shiftcode/sc-ng-commons-public/compare/@shiftcode/ngx-core@1.0.0...@shiftcode/ngx-core@1.1.0) (2022-06-14)
+
+
+### Features
+
+* **is-input-element:** new helper function ([faebe03](https://github.com/shiftcode/sc-ng-commons-public/commit/faebe03eafccd9cfbc144c8d731c5050d5a4611a))
+
+
+
+
+
 # 1.0.0 (2022-01-22)
 
 **Note:** Version bump only for package @shiftcode/ngx-core